% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPheno.R
\name{bernoulliscaling}
\alias{bernoulliscaling}
\title{Sacling factore for geno inorder to get the expected heritability 
from a Bernoulli distribution}
\usage{
bernoulliscaling(geno, h2 = 0.5)
}
\arguments{
\item{geno}{Vector of genetic values.}

\item{h2}{Expected heritability.}
}
\description{
Sacling factore for geno inorder to get the expected heritability 
from a Bernoulli distribution
}
\keyword{internal}

