% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genepos}
\alias{genepos}
\title{Gene positions of Arabidopsis thaliana.}
\format{A data frame:
\describe{
  \item{chr}{Chromosomes were genes are located}
  \item{pos}{Position at the chromosomes were genes are located}
}}
\source{
Data are avalebel from: 
\url{ftp://ftp.arabidopsis.org/home/tair/Genes/TAIR10_genome_release/TAIR10_gff3/TAIR10_GFF3_genes.gff}
}
\usage{
genepos
}
\description{
A dataset of of gene positions of Arabidopsis thaliana genes.
}
\keyword{datasets}

