% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPheno.R
\name{normalPheno}
\alias{normalPheno}
\title{Simulates a normal distributed phenotype with effects either normal 
or gamma distributed.}
\usage{
normalPheno(snp, n.rep = 1, n.loci = 50, h2 = 0.5, snp.probs = NULL,
  eff.dist = c("gaussian", "gamma"), shape = 0.5)
}
\arguments{
\item{snp}{An object of S4 class \linkS4class{snpData}.}

\item{n.rep}{Number of replication per individual. For the replication no 
effects at the phenotype are simulated.}

\item{n.loci}{Number of loci with effect at the phenotype.}

\item{h2}{Trait heritability.}

\item{snp.probs}{A verctor of probertibilities for each SNP to be selected 
as a causal loci.}

\item{eff.dist}{Distribution from which the effects are drown.}

\item{shape}{See \code{\link[stats]{rgamma}} (only considert if 'eff.dist = 
gamma').}
}
\description{
Simulates a normal distributed phenotype with effects either normal 
or gamma distributed.
}

