% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtcat.dataPackage.R
\docType{package}
\name{qtcat-package}
\alias{qtcat}
\alias{qtcat-package}
\title{Datasets for the Quantitative Trait Cluster Association Test}
\description{
Data sets and simulation tools for QTCAT.
}
\author{
Jonas Klasen
}

