% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPheno.R
\name{rres}
\alias{rres}
\title{Random residuals which match the expected heritability}
\usage{
rres(n, Va = 1, h2 = 0.5)
}
\arguments{
\item{n}{Number of residuals.}

\item{Va}{Genetic variance.}

\item{h2}{Expected heritability.}
}
\description{
Random residuals which match the expected heritability
}
\keyword{internal}

