% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{snp}
\alias{snp}
\title{An Arabidopsis thaliana SNP data object.}
\format{snpMatrix object:
\describe{
  \item{snp}{An object of S4 class \linkS4class{snpMatrix}}
}}
\source{
Horton, Matthew W., Angela M. Hancock, Yu S. Huang, Christopher 
Toomajian, Susanna Atwell, Adam Auton, N. Wayan Muliyati, et al. 2012. 
Genome-Wide Patterns of Genetic Variation in Worldwide Arabidopsis thaliana 
Accessions from the RegMap Panel. Nature Genetics 44 (2): 212-16. 
\url{http://doi.org/10.1038/ng.1042.} 

Data are available from: 
\url{http://bergelson.uchicago.edu/wp-content/uploads/2015/04/call_method_75.tar.gz}
}
\usage{
snp
}
\description{
A dataset of 1307 Arabidopsis thaliana accessions genotyped and 214051 SNPs. 
The data required 285 Mb of memory only to be loaded, so make sure your 
memory is big enough.
}
\keyword{datasets}

