% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{snpclust}
\alias{snpclust}
\title{A SNP cluster object of the Arabidopsis thaliana data.}
\format{Clustering object:
\describe{
  \item{snpclust}{A clustering of all SNPs}
}}
\usage{
snpclust
}
\description{
A SNP cluster object of the Arabidopsis thaliana data.
}
\keyword{datasets}

