% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtcat.dataPackage.R
\docType{package}
\name{qtcat-package}
\alias{qtcat}
\alias{qtcat-package}
\title{Datasets for the Quantitative Trait Cluster Association Test}
\description{
A data set and simulation tools for QTCAT.
    Attention: the package contains a large genetic data set, in order to analyse it make 
    sure you have at least 8 GB free RAM.
}
\author{
Jonas Klasen
}

