% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPheno.R
\name{snpProb}
\alias{snpProb}
\title{Estimate sample probability for each SNP from gene density}
\usage{
snpProb(gene.pos, snp.pos, bw = 10000)
}
\arguments{
\item{gene.pos}{a matrix of two rows. The first row contains the 
chromosome, the second row contains the position of a gene (e.g. start in bp).}

\item{snp.pos}{a matrix of two rows. the first row contains the 
chromosome, the second row contains the position of a SNP  bp.}

\item{bw}{see \code{\link[stats]{density}}.}
}
\description{
Estimate sample probability for each SNP from gene density
}

